+++ modules/template/elements/FileDownloadElement.php
@@ -163,7 +163,7 @@ class FileDownloadElement extends BaseElementContent implements \Stringable
     {
         $id = 'fileDownloadElement-' . $this->id;
 
-        $result = $form->field($this, 'title')
+        return $form->field($this, 'title')
             . $form->field($this, 'showFileinfo')->checkbox()
             . $form->field($this, 'showIcon')->checkbox()
             . Html::beginTag('div', ['id' => $id, 'class' => 'file-upload-container clearfix'])
@@ -188,18 +188,25 @@ class FileDownloadElement extends BaseElementContent implements \Stringable
                     'id' => $id . '-progress',
                     'options' => ['style' => 'display:block;margin-left:150px;width:500px'],
                 ])
-            . Html::endTag('div');
+            . Html::endTag('div')
+            . $this->renderCollapsableEditForm($form);
+    }
 
+    private function renderCollapsableEditForm(ActiveForm $form): string
+    {
         ob_start();
+
         CollapsableFormGroup::begin([
             'defaultState' => false,
             'label' => Yii::t('CustomPagesModule.base', 'Advanced'),
         ]);
+
         echo $form->field($this, 'style')
             . $form->field($this, 'cssClass')
             . $form->field($this, 'file_guid')->hiddenInput(['class' => 'file-guid'])->label(false);
+
         CollapsableFormGroup::end();
 
-        return $result . ob_get_flush();
+        return ob_get_clean();
     }
 }